Refactor complete view (#59)
* refactor create

* try to fix last codeclimate smell

* restructure files
